wrapped up commit
The current pull request consists of a working simulation where an actor
owning an instance of btcwallet successfully connects to btcd, creates an
encrypted wallet of 1000 addresses and since it has no fundsm it normally
exits. Some refactoring of the initial code has been made where it seemed
to be appropriate. The process of actor spawning has been designed in a
way that it can scale to create more actors without a problem.