Connect wallet-btcd, create wallet addresses, exit
A working simulation is possible now where an actor owning
an instance of btcwallet successfully connects to btcd,
creates an encrypted wallet of 1000 addresses and since it
has no funds, the actor normally exits. Some refactoring of
the initial code has been made where it seemed to be appropriate.
The process of actor spawning has been designed in a way that
it can scale to create more actors without a problem. Finally
a .gitignore file has been added.