Add mining
Mining has been added with the use of the NewMiner function so now a
complete simulation is possible to happen. The way NewMiner works is
that it fires up a fresh btcd instance with --generate and --miningaddr
options enabled (--miningaddr is specified as many times as the number
of actors since every actor uses one of its addresses to collect mining
funds) to mine coins.

Also the communication between actors and the main goroutine has been
upgraded to accomodate actors with insufficient funds.

Finally some refactoring was done in favor of eliminating almost all
globals.