Declutter logging output and block until coinbase matures
Logs were really messed up with stuff we really don't care about such as
TXST notifications or OnBlockConnected notifications piled up at start.
Also it's more convenient to block actors before entering the select
statement waiting for a matured coinbase.
Another thing to note is that in SendFromMinConf while using one satoshi
as a sending amount, the rest of the UTXO (50 BTC) gets back to the
coinbase as mining reward, something we don't want to happen.