Calculate avg txn per second
Every actor after quiting sending/receiving transactions, lists its
latest transactions based on a fixed number and sends them back to
main. Main saves them all in a slice and with the help of an extra map,
it caclulates the average transactions per second.

a.WaitForShutdown has been replaced with the defered call to a.wg.Wait at
the start of a.Start and some fixes regarding sending to a.stop were
made, so a simulation can normally end.