Update for recent chainhash-related API changes.
This updates all code to make use of the new chainhash package since the
old wire.ShaHash type and related functions have been removed in favor
of the abstracted package.

Also, update all of the license headers to be consistent with the
btcsuite code base which includes relicensing it to the btcsuite
developers.