TravisCI: Update to latest Go versions.
The vet tool moved into the Go source tree as of Go 1.5. Its previous
location in the x/tools repo was deprecated at that time and has now
been removed.

This updates the .travis.yml configuration to avoid fetching vet from
the old location and to simply use the version now available as part of
the standard Go install.

Also, while here, remove the check for changing to tool path since it is
no longer needed.